+++ src/components/BarChart/index.tsx
@@ -14,6 +14,8 @@ import { Point } from "../Point";
 import { Container } from "../Container";
 import { CustomTooltipProps, IChart } from "../LineChart";
 
+import useBarChartState from "./useBarChartState";
+
 export interface ChartSeriesType {
   name: string;
   data: number[];
@@ -22,6 +24,7 @@ export interface ChartSeriesType {
 export interface IBarChart extends IChart {
   data: string[];
   series: ChartSeriesType[];
+  displayDirection?: "vertical" | "horizontal";
 }
 
 const CustomTooltip = ({
@@ -66,53 +69,47 @@ const CustomTooltip = ({
   );
 };
 
-const getHorizontalCoordinates = (max: number, gap: number) => {
-  const horizontalLines = [];
-
-  for (let index = 0; gap * index <= max; index++) {
-    horizontalLines.push(gap * index + 5);
-  }
-
-  return horizontalLines;
-};
-
 export const BarChart = ({
   data,
   series,
   colors,
   fontFamily,
   formatValue = (value) => value.toLocaleString("en-US"),
+  displayDirection = "horizontal",
 }: IBarChart) => {
-  const chartData: any = data.map((name) => ({ name }));
-  const keyNames: string[] = [];
-
-  series.forEach((element) => {
-    keyNames.push(element.name);
-    element.data.forEach((item, index) => {
-      if (chartData[index]) {
-        chartData[index][element.name] = item;
-      }
+  const [{ chartData, keyNames }, { isHide, setHide, getCoordinates }] =
+    useBarChartState({
+      data,
+      series,
     });
-  });
 
   return (
     <div className="bar-chart-container">
       <div className="bar-chart-graphic-container">
         <ResponsiveContainer width="100%" height={120 * data.length}>
-          <RechartsBarChart data={chartData} layout="vertical" barGap={8}>
-            <CartesianGrid
-              strokeDasharray="12 12"
-              horizontalCoordinatesGenerator={(props) =>
-                getHorizontalCoordinates(
-                  props.yAxis.height,
-                  props.yAxis.bandSize
-                )
-              }
-            />
+          <RechartsBarChart
+            data={chartData}
+            layout={displayDirection === "vertical" ? "horizontal" : "vertical"}
+            barGap={8}
+          >
+            {displayDirection === "horizontal" ? (
+              <CartesianGrid
+                strokeDasharray="12 12"
+                horizontalCoordinatesGenerator={(props) =>
+                  getCoordinates(props.yAxis.height, props.yAxis.bandSize)
+                }
+              />
+            ) : (
+              <CartesianGrid
+                strokeDasharray="12 12"
+                verticalCoordinatesGenerator={(props) =>
+                  getCoordinates(props.xAxis.width, props.xAxis.bandSize)
+                }
+              />
+            )}
             <XAxis
-              type="number"
+              type={displayDirection === "vertical" ? "category" : "number"}
               axisLine={false}
-              fontFamily={fontFamily}
               tick={(props) => (
                 <text
                   x={props.x + 10}
@@ -123,12 +120,14 @@ export const BarChart = ({
                   fontWeight={500}
                   fontFamily={fontFamily}
                 >
-                  {formatValue(props.payload.value)}
+                  {displayDirection === "horizontal"
+                    ? formatValue(props.payload.value)
+                    : chartData[props.payload.index].name}
                 </text>
               )}
             />
             <YAxis
-              type="category"
+              type={displayDirection === "vertical" ? "number" : "category"}
               axisLine={false}
               tick={(props) => (
                 <text
@@ -140,12 +139,15 @@ export const BarChart = ({
                   fontWeight={500}
                   fontFamily={fontFamily}
                 >
-                  {chartData[props.payload.index].name}
+                  {displayDirection === "horizontal"
+                    ? chartData[props.payload.index].name
+                    : formatValue(props.payload.value)}
                 </text>
               )}
             />
-            {keyNames.map((name, index) => (
+            {keyNames?.map((name, index) => (
               <Bar
+                hide={isHide?.(name)}
                 key={`bar-${name}-${index}`}
                 dataKey={name}
                 fill={colors[index % colors.length]}
@@ -167,12 +169,21 @@ export const BarChart = ({
         </ResponsiveContainer>
       </div>
       <div className={clsx("flex-center", "bar-chart-legend-container")}>
-        {keyNames.map((name, index) => (
+        {keyNames?.map((name, index) => (
           <span
             key={`bar-char-legend-${index}`}
             className={clsx("flex-center", "variant-body1")}
             style={{ fontFamily }}
           >
+            <input
+              className="checkbox"
+              type="checkbox"
+              checked={!isHide?.(name)}
+              onChange={(e) => {
+                setHide?.((prev) => ({ ...prev, [name]: !e.target.checked }));
+              }}
+              style={{ accentColor: colors[index % colors.length] }}
+            />
             <Point color={colors[index % colors.length]} /> {name}{" "}
           </span>
         ))}